+++ NEWS.md
@@ -1,5 +1,9 @@
 # v 0.4.2.00x (dev version)
 
+## Major changes
+
+- All graphs are now constructed with distances calculated using the "cheap" measure of `geodist::geodist()` is maximal graph distances are < 100km, otherwise with "haversine". No distances are calculated with geodesic distances.
+
 ## Minor changes
 
 - "geodist" package is now a hard dependency ("Imports" not "Suggests").